Fixed build dependencies
Added sources for Go package golang.org/x/oauth2 to the RPM spec file.
Started moving generic utility functions to a separate Go package.
Wrote a short function to avoid depencency on github.com/tidwall/gjson,
mostly because of how 3rd party Go packages complicate RPM building.

Work in progress on #28.
